requirements.txt
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
python-dotenv==0.21.1
Faker==18.3.1
flake8==5.0.4
psycopg2-binary==2.9.5
flask-restful==0.3.9
marshmallow-sqlalchemy==0.29.0
flask-jwt-extended==4.4.4
Flask-Caching==2.0.2
cachelib==0.9.0
